feat: add css-utils module and refactor preprocessCss function
