+++ lib/prompts.mjs
@@ -3,16 +3,11 @@
 // and by the CLI in bin/mint-ds.mjs. Keep this file dependency-free so it runs
 // in plain Node without a build step.
 
+import { preprocessCss } from './css-utils.mjs'
+export { preprocessCss }
+
 const MODEL = 'claude-sonnet-4-20250514'
 const ANTHROPIC_URL = 'https://api.anthropic.com/v1/messages'
-
-export function preprocessCss(raw) {
-  return String(raw)
-    .replace(/\/\*[\s\S]*?\*\//g, ' ')  // block comments
-    .replace(/\/\/[^\n]*/g, ' ')         // line comments (SCSS/LESS)
-    .replace(/\s+/g, ' ')
-    .trim()
-}
 
 export const AUDIT_SYSTEM_PROMPT = `You are an expert design-system auditor with deep knowledge of CSS architecture, design tokens, and visual consistency. Your role is to analyze CSS codebases, identify color, typography, and spacing patterns, and surface actionable data for design system extraction.
